Folder Restructure
-changed up the folder structure
- did minor changes
