+++ Layouts/Navigation/NavLinkItem.razor.css
@@ -74,6 +74,7 @@
 }
 
 .nav-item__label--button:hover {
+	color: var(--navitem-color-lightest);
 	background-color: var(--navitem-color-light);
 	border: 2px solid var(--navitem-color-light);
 }
