+++ Layouts/Navigation/SideMenu.razor.css
