+++ Layouts/UI/SearchInput.razor
@@ -1,5 +1,3 @@
-@using System.Timers
-
 <div class="search-input">
 	<input class="search-input__field" type="text" value="@_inputValue" placeholder="@Placeholder"
 		@oninput="HandleInput" />
