--- Pages/AboutMe.razor
@@ -1,29 +1,0 @@
-@page "/aboutme"
-
-<div class="about-me">
-    <div class="about-me__content">
-        <div class="about-me__profile">
-            <img class="about-me__profile-img" src="@AppContent.User.ProfileImageUrl"
-                alt="@AppContent.User.ProfileImageUrl" />
-
-            <div class="about-me__social-icons">
-                @foreach (var icon in AppContent.User.SocialLinks)
-                {
-                    <a class="about-me__social-link" href="@icon.Detail" target="_blank" rel="noopener noreferrer"
-                        title="@icon.Title">
-                        <img class="about-me__social-icon" src="@icon.Title" alt="@icon.Title" />
-                    </a>
-                }
-            </div>
-        </div>
-
-        <div class="about-me__details">
-            <h3 class="about-me__title">@AppContent.User.AboutMeTitle</h3>
-            <p class="about-me__description">@AppContent.User.AboutMeDescription</p>
-        </div>
-    </div>
-</div>
-
-@code {
-    [Inject] private AppContent AppContent { get; set; } = null!;
-}
