--- Pages/AboutMe.razor.css
@@ -1,163 +1,0 @@
-.about-me {
-	--section-padding: 3rem 1rem 0rem;
-	--content-gap: 2rem;
-	--profile-padding: 1rem;
-	--profile-img-width: 220px;
-	--profile-img-height: 250px;
-	--social-icon-size: 1rem;
-	--social-icon-gap: 1.5rem;
-	--title-font-size: 1.25rem;
-	--title-line-height: 1.4em;
-	--title-margin-bottom: 0.8rem;
-	--desc-font-size: 0.95rem;
-	--desc-line-height: 1.6em;
-}
-
-.about-me {
-	background-color: var(--color-background-main);
-	/* White background */
-	color: var(--color-primary-main);
-	/* Black text */
-	padding: var(--section-padding);
-	min-height: 500px;
-	display: flex;
-	justify-content: center;
-}
-
-.about-me__content {
-	gap: var(--content-gap);
-	width: 100%;
-	display: flex;
-	flex-wrap: wrap;
-	flex-direction: row;
-	align-items: center;
-	justify-content: center;
-	height: 100%;
-	border-bottom: 1px solid var(--color-primary-muted);
-	/* Soft separator */
-	margin-bottom: 2rem;
-}
-
-.about-me__profile {
-	padding: var(--profile-padding);
-	background-color: var(--color-primary-muted);
-	/* Subtle background */
-	border: 1px solid var(--color-primary-muted);
-	width: 200px;
-	display: flex;
-	align-items: center;
-	flex-direction: column;
-}
-
-.about-me__profile-img {
-	width: var(--profile-img-width);
-	height: var(--profile-img-height);
-	background-color: white;
-	object-fit: cover;
-	border-radius: 8px;
-	/* Slight round edges for modern look */
-}
-
-.about-me__social-icons {
-	gap: var(--social-icon-gap);
-	margin-top: 1rem;
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: center;
-}
-
-.about-me__social-icon {
-	width: var(--social-icon-size);
-	height: var(--social-icon-size);
-	object-fit: contain;
-	transition: transform 0.2s ease-out;
-}
-
-.about-me__social-icon:hover {
-	transform: scale(1.1);
-}
-
-.about-me__details {
-	min-width: 300px;
-	flex: 1;
-	flex-direction: column;
-	align-items: center;
-	text-align: center;
-}
-
-.about-me__title {
-	font-size: var(--title-font-size);
-	margin-bottom: var(--title-margin-bottom);
-	line-height: var(--title-line-height);
-	color: var(--color-primary-main);
-	/* Black text for title */
-}
-
-.about-me__description {
-	font-size: var(--desc-font-size);
-	line-height: var(--desc-line-height);
-	color: var(--color-primary-muted);
-	/* Muted gray for description */
-}
-
-@media (min-width: 576px) {
-	.about-me {
-		--section-padding: 3.5rem 1.5rem 0rem;
-		--profile-img-width: 280px;
-		--profile-img-height: 360px;
-		--desc-font-size: 1rem;
-	}
-
-	.about-me__profile {
-		width: 250px;
-	}
-}
-
-@media (min-width: 768px) {
-	.about-me {
-		--section-padding: 4rem 2rem 0rem;
-		--content-gap: 3rem;
-		--profile-img-width: 220px;
-		--profile-img-height: 280px;
-		--title-font-size: 1.4rem;
-		--title-margin-bottom: 1rem;
-	}
-
-	.about-me__content {
-		align-items: stretch;
-		justify-content: flex-start;
-		height: auto;
-	}
-
-	.about-me__profile {
-		width: 200px;
-		margin-bottom: 2rem;
-	}
-
-	.about-me__details {
-		display: flex;
-		flex-direction: column;
-		flex: 1;
-		align-items: center;
-		justify-content: center;
-		text-align: start;
-		min-height: 100%;
-	}
-}
-
-@media (min-width: 992px) {
-	.about-me {
-		--section-padding: 5rem 3rem 0rem;
-		--social-icon-size: 1.2rem;
-		--title-font-size: 1.6rem;
-	}
-}
-
-@media (min-width: 1200px) {
-	.about-me {
-		--section-padding: 6rem 4rem 0rem;
-		--social-icon-size: 1.3rem;
-		--social-icon-gap: 2rem;
-		--title-font-size: 1.8rem;
-	}
-}
