+++ Pages/Contact.razor.css
@@ -4,7 +4,7 @@
 
 .contact {
 	padding: var(--contact-page-padding);
-	overflow-y: hidden;
+	overflow: hidden;
 }
 
 .contact__grid {
@@ -13,10 +13,12 @@
 	gap: 5rem;
 }
 
-.contact__info {
-	display: flex;
-	flex-direction: column;
-	gap: 2rem;
+.contact__info-grid {
+	display: grid;
+	gap: 1.5rem;
+	grid-template-columns: 1fr;
+	grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+	margin-bottom: 4rem;
 }
 
 @media (min-width: 768px) {
@@ -34,5 +36,9 @@
 
 	.contact__info {
 		justify-content: center;
+	}
+
+	.contact__info-grid {
+		margin-bottom: 0;
 	}
 }
