--- Pages/Skills.razor
@@ -1,54 +1,0 @@
-@page "/skills"
-
-<div class="skills">
-	<h2 class="skills__title">Skills</h2>
-
-	<div class="skills__grid">
-		<div class="skills__group">
-			<h3 class="skills__group-title">Game Development</h3>
-			<ul class="skills__list">
-				<li class="skills__item">
-					<img src="https://cdn.jsdelivr.net/gh/devicons/devicon/icons/csharp/csharp-original.svg" alt="C#"
-						class="skills__icon" />
-					C#
-				</li>
-				<li class="skills__item">
-					<img src="https://cdn.jsdelivr.net/gh/devicons/devicon/icons/godot/godot-original.svg" alt="Godot"
-						class="skills__icon" />
-					Godot
-				</li>
-				<li class="skills__item">
-					<img src="https://cdn.jsdelivr.net/gh/devicons/devicon/icons/unity/unity-original.svg" alt="Unity"
-						class="skills__icon" />
-					Unity
-				</li>
-			</ul>
-		</div>
-
-		<div class="skills__group">
-			<h3 class="skills__group-title">Web Development</h3>
-			<ul class="skills__list">
-				<li class="skills__item">
-					<img src="https://cdn.jsdelivr.net/gh/devicons/devicon/icons/dot-net/dot-net-original.svg"
-						alt=".NET" class="skills__icon" />
-					.NET
-				</li>
-				<li class="skills__item">
-					<img src="https://cdn.jsdelivr.net/gh/devicons/devicon@latest/icons/blazor/blazor-original.svg"
-						alt="Blazor" class="skills__icon" />
-					Blazor
-				</li>
-				<li class="skills__item">
-					<img src="https://cdn.jsdelivr.net/gh/devicons/devicon/icons/html5/html5-original.svg" alt="HTML5"
-						class="skills__icon" />
-					HTML
-				</li>
-				<li class="skills__item">
-					<img src="https://cdn.jsdelivr.net/gh/devicons/devicon/icons/css3/css3-original.svg" alt="CSS3"
-						class="skills__icon" />
-					CSS
-				</li>
-			</ul>
-		</div>
-	</div>
-</div>
