--- Pages/Skills.razor.css
@@ -1,65 +1,0 @@
-.skills {
-	padding: 2rem 1rem;
-	background-color: var(--color-background-main);
-	color: var(--color-primary-main);
-	font-family: var(--font-family-body);
-}
-
-.skills__title {
-	font-family: var(--font-family-heading);
-	font-size: 1.75rem;
-	margin-bottom: 2rem;
-	text-align: center;
-}
-
-.skills__grid {
-	display: grid;
-	grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-	gap: 1.5rem;
-}
-
-.skills__group {
-	border: 1px solid var(--color-background-muted);
-	padding: 1.5rem;
-	display: flex;
-	flex-direction: column;
-	gap: 1rem;
-}
-
-.skills__group-title {
-	color: black;
-	font-size: 1.125rem;
-	font-family: var(--font-family-heading);
-	border-bottom: 1px solid var(--color-background-muted);
-	padding-bottom: 0.5rem;
-	margin-bottom: 0.5rem;
-}
-
-.skills__list {
-	display: grid;
-	grid-template-columns: 1fr 1fr;
-	gap: 0.75rem 1rem;
-	list-style: none;
-	padding: 0;
-	margin: 0;
-}
-
-.skills__item {
-	display: flex;
-	align-items: center;
-	gap: 0.5rem;
-	font-size: 0.95rem;
-	color: var(--color-primary-main);
-}
-
-.skills__icon {
-	width: 20px;
-	height: 20px;
-	object-fit: contain;
-	filter: grayscale(1) brightness(0.8);
-	transition: filter 0.2s ease-in-out;
-}
-
-.skills__icon:hover {
-	filter: none;
-}
