+++ Program.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using MyPortfolio;
-using MyPortfolio.Models;
+using MyPortfolio.Models.Profile;
 using MyPortfolio.Services.State;
 
 WebAssemblyHostBuilder builder = WebAssemblyHostBuilder.CreateDefault(args);
