+++ _Imports.razor
@@ -1,23 +1,18 @@
-﻿@using System.Net.Http
-@using System.Net.Http.Json
-@using Microsoft.AspNetCore.Components.Forms
+﻿@using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components.Web.Virtualization
-@using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using MyPortfolio
-@using MyPortfolio.Contracts
+@using System.Text.Json
+@using System.Timers
 @using MyPortfolio.Helpers.Contact
 @using MyPortfolio.Helpers.Navigation
-@using MyPortfolio.Layout
-@using MyPortfolio.Layout.Navigation
-@using MyPortfolio.Layout.UI
-@using MyPortfolio.Models
+@using MyPortfolio.Layouts
+@using MyPortfolio.Layouts.Navigation
+@using MyPortfolio.Layouts.UI
 @using MyPortfolio.Models.Contact
 @using MyPortfolio.Models.Home
+@using MyPortfolio.Models.Profile
 @using MyPortfolio.Models.Service
 @using MyPortfolio.Models.Navigation
 @using MyPortfolio.Models.UI
-@using MyPortfolio.Pages
 @using MyPortfolio.Services.State
